# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  garden.CreateGardenRequest:
    properties:
      description:
        example: A beautiful garden for growing vegetables
        type: string
      name:
        example: My First Garden
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  garden.PlantRequest:
    properties:
      plant_type_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      position:
        example: 0
        maximum: 8
        minimum: 0
        type: integer
    required:
    - plant_type_id
    - position
    type: object
...
  garden.UpdateGardenRequest:
    properties:
      description:
        example: Updated garden description
        type: string
      name:
        example: Updated Garden Name
        type: string
    type: object
...
info:
  contact: {}
  description: A robust RESTful API for a virtual garden management game built with
    Go.
  title: Virtual Garden Management System API
  version: "1.0"
paths:
...
  /gardens:
    get:
      consumes:
      - application/json
      description: Get all gardens for the current user
      produces:
      - application/json
      responses:
        "200":
          description: List of user gardens
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get user gardens
      tags:
      - gardens
    post:
      consumes:
      - application/json
      description: Create a new garden for the current user
      parameters:
      - description: Garden creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/garden.CreateGardenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created garden
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Create a new garden
      tags:
      - gardens
  /gardens/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a garden and all its plants
      parameters:
      - description: Garden ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Garden deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request - Invalid garden ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Delete garden
      tags:
      - gardens
    get:
      consumes:
      - application/json
      description: Get a specific garden with all its plants
      parameters:
      - description: Garden ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Garden details with plants
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request - Invalid garden ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Get garden details
      tags:
      - gardens
    put:
      consumes:
      - application/json
      description: Update garden information
      parameters:
      - description: Garden ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: id
        required: true
        type: string
      - description: Garden update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/garden.UpdateGardenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated garden
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Update garden
      tags:
      - gardens
  /gardens/{id}/plants:
    post:
      consumes:
      - application/json
      description: Plant a seed in a garden
      parameters:
      - description: Garden ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: id
        required: true
        type: string
      - description: Plant data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/garden.PlantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Planted seed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Garden or plant type not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Position already occupied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Plant a seed
      tags:
      - plants
  /gardens/{id}/plants/{plantId}:
    delete:
      consumes:
      - application/json
      description: Remove a plant from the garden
      parameters:
      - description: Garden ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: id
        required: true
        type: string
      - description: Plant ID
        example: '"123e4567-e89b-12d3-a456-426614174001"'
        in: path
        name: plantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Plant removed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Plant not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Remove a plant
      tags:
      - plants
  /gardens/{id}/plants/{plantId}/harvest:
    post:
      consumes:
      - application/json
      description: Harvest a mature plant to get rewards
      parameters:
      - description: Garden ID
        example: '"123e4567-e89b-12d3-a456-426614174000"'
        in: path
        name: id
        required: true
        type: string
      - description: Plant ID
        example: '"123e4567-e89b-12d3-a456-426614174001"'
        in: path
        name: plantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Harvest results
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request - Plant not ready for harvest
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Plant not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - bearer: []
      summary: Harvest a plant
      tags:
      - plants
  /plants:
    get:
      consumes:
      - application/json
      description: Get a list of all available plant types
      produces:
      - application/json
      responses:
        "200":
          description: List of plant types
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all plant types
      tags:
      - plants
...
securityDefinitions:
  bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"